refactor(ui): split five 1000+ line screens into focused files
Every Compose screen in the app had grown the same way: the screen
composable, the components it happens to use, the dialogs it raises, and
the drawing helpers those need, all in one file. Split along the seams
that were already there — no logic moved, every composable is unchanged
where it now sits.

  ClassTableScreen.kt   1647 -> 525 + Grid 194 + CourseCell 143
                                    + ConflictCells 421 + DetailDialog 255
  OnboardingScreen.kt   1113 -> 509 + Components 281 + Icons 190
  SettingsScreen.kt     1112 -> 615 + Rows 307 + Dialogs 71
  HomeScreen.kt         1027 -> 387 + Sections 250 + AssignmentList 257
  ScoreScreen.kt        1000 -> 224 + TrendChart 440 + CourseList 160
                                    + CourseDetailDialog 58

Kotlin's `private` is file-scoped like Swift's, so composables now called
across a file boundary widened to `internal`. Nothing became public.

The split boundaries are meant to be meaningful, not just size-driven:
衝堂 cells are separate from the plain course cell because dividing one
grid cell between courses and keeping each half tappable is a different
layout problem; the score trend chart is separate because monotone-cubic
path fitting and axis-domain maths have nothing to do with listing
grades; onboarding's Canvas illustrations are separate from its flow.

SettingsScreen.kt is still 615 lines, almost all of it one composable.
That one wants section extraction rather than file surgery, and is left
for a follow-up so this commit stays a pure move.
