+++ app/src/main/java/org/ntust/app/tigerduck/ui/screen/home/HomeScreen.kt
@@ -1,9 +1,10 @@
+// The Home tab: greeting, the reorderable section list, and the two
+// dialogs it owns. Section bodies are in HomeSections.kt and the
+// assignment list in HomeAssignmentList.kt.
+
 package org.ntust.app.tigerduck.ui.screen.home
 
-import android.content.Context
-import android.content.Intent
 import androidx.activity.compose.BackHandler
-import androidx.compose.animation.core.Animatable
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.animation.core.spring
 import androidx.compose.animation.core.tween
@@ -16,7 +17,6 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -25,33 +25,14 @@ import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.lazy.itemsIndexed
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.Undo
 import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.Check
-import androidx.compose.material.icons.filled.CheckCircle
-import androidx.compose.material.icons.filled.Close
-import androidx.compose.material.icons.filled.DragHandle
-import androidx.compose.material.icons.filled.Inbox
-import androidx.compose.material.icons.filled.Lock
-import androidx.compose.material.icons.filled.VisibilityOff
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.SegmentedButton
-import androidx.compose.material3.SegmentedButtonColors
-import androidx.compose.material3.SegmentedButtonDefaults
-import androidx.compose.material3.SingleChoiceSegmentedButtonRow
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.Text
@@ -75,18 +56,13 @@ import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.draw.scale
 import androidx.compose.ui.draw.shadow
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.lerp
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalDensity
-import androidx.compose.ui.platform.LocalLayoutDirection
 import androidx.compose.ui.platform.LocalResources
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.IntOffset
-import androidx.compose.ui.unit.LayoutDirection
 import androidx.compose.ui.unit.dp
 import androidx.core.net.toUri
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
@@ -99,24 +75,19 @@ import kotlinx.coroutines.launch
 import org.ntust.app.tigerduck.AppConstants
 import org.ntust.app.tigerduck.R
 import org.ntust.app.tigerduck.data.model.Assignment
-import org.ntust.app.tigerduck.data.model.AssignmentFilter
 import org.ntust.app.tigerduck.shared.Course
 import org.ntust.app.tigerduck.data.model.HomeSection
 import org.ntust.app.tigerduck.shared.clock.AppClock
 import org.ntust.app.tigerduck.ui.AppState
-import org.ntust.app.tigerduck.ui.component.AssignmentItem
 import org.ntust.app.tigerduck.ui.component.ComingSoonDialog
 import org.ntust.app.tigerduck.ui.component.TigerDuckDialog
-import org.ntust.app.tigerduck.ui.component.EmptyStateView
 import org.ntust.app.tigerduck.ui.component.PageHeader
-import org.ntust.app.tigerduck.ui.component.SectionHeader
 import org.ntust.app.tigerduck.ui.component.ServerKind
 import org.ntust.app.tigerduck.ui.component.ServerStatusIcons
 import org.ntust.app.tigerduck.ui.component.SyncIndicator
 import org.ntust.app.tigerduck.ui.component.TigerPullToRefresh
 import org.ntust.app.tigerduck.ui.component.rememberAppClockVersion
 import org.ntust.app.tigerduck.ui.navigation.icon
-import org.ntust.app.tigerduck.ui.theme.TigerDuckTheme
 import java.util.Calendar
 import kotlin.math.abs
 import kotlin.math.roundToInt
@@ -408,259 +379,16 @@ fun HomeScreen(
         )
     }
 }
-
 @Composable
-private fun ReorderableSection(
-    section: HomeSection,
-    isEditing: Boolean,
-    isDragging: Boolean,
-    dragOffsetY: Float,
-    onStartEditing: () -> Unit,
-    onDragStart: () -> Unit,
-    onDrag: (Float) -> Unit,
-    onDragEnd: () -> Unit,
-    onMeasured: (height: Float) -> Unit,
-    onRemove: () -> Unit,
-    content: @Composable () -> Unit,
-) {
-    // Material-style elevation ramp for the dragging card — no wiggle.
-    val elevation by animateFloatAsState(
-        targetValue = if (isDragging) 12f else 0f,
-        label = "drag-elevation",
-    )
-
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .onGloballyPositioned { coords ->
-                onMeasured(coords.size.height.toFloat())
-            }
-            .offset { IntOffset(0, dragOffsetY.roundToInt()) }
-            .then(
-                if (isDragging) Modifier.shadow(elevation.dp, RoundedCornerShape(18.dp))
-                else Modifier
-            )
-            .then(
-                // Only long-press-to-enter-edit-mode lives here. Reordering
-                // drags come from the dedicated handle below, so tap events
-                // on child cards are never stolen by this gesture detector.
-                if (!isEditing) {
-                    Modifier.pointerInput(section.id) {
-                        detectTapGestures(onLongPress = { onStartEditing() })
-                    }
-                } else {
-                    Modifier
-                }
-            )
-    ) {
-        content()
-
-        if (isEditing) {
-            // Top-right row: big touch-friendly drag handle + delete.
-            Row(
-                modifier = Modifier
-                    .align(Alignment.TopEnd)
-                    .padding(end = 20.dp, top = 4.dp),
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.spacedBy(8.dp),
-            ) {
-                // Drag handle — 44dp hit target, starts reorder on first
-                // touch-and-move (no long-press required).
-                Box(
-                    modifier = Modifier
-                        .size(44.dp)
-                        .clip(CircleShape)
-                        .background(MaterialTheme.colorScheme.surfaceVariant)
-                        .pointerInput(section.id) {
-                            detectDragGestures(
-                                onDragStart = { onDragStart() },
-                                onDragEnd = { onDragEnd() },
-                                onDragCancel = { onDragEnd() },
-                                onDrag = { change, drag ->
-                                    change.consume()
-                                    onDrag(drag.y)
-                                },
-                            )
-                        },
-                    contentAlignment = Alignment.Center,
-                ) {
-                    Icon(
-                        Icons.Filled.DragHandle,
-                        contentDescription = stringResource(R.string.tab_editor_drag_reorder),
-                        tint = MaterialTheme.colorScheme.onSurfaceVariant,
-                        modifier = Modifier.size(24.dp),
-                    )
-                }
-                // Delete
-                Box(
-                    modifier = Modifier
-                        .size(44.dp)
-                        .clip(CircleShape)
-                        .background(MaterialTheme.colorScheme.error)
-                        .clickable(enabled = !isDragging) { onRemove() },
-                    contentAlignment = Alignment.Center,
-                ) {
-                    Icon(
-                        Icons.Filled.Close,
-                        contentDescription = stringResource(R.string.action_remove),
-                        tint = MaterialTheme.colorScheme.onError,
-                        modifier = Modifier.size(24.dp),
-                    )
-                }
-            }
-        }
+private fun greetingText(): String {
+    val hour = AppClock.calendar().get(Calendar.HOUR_OF_DAY)
+    return when {
+        hour < 6 -> stringResource(R.string.greeting_very_early)
+        hour < 12 -> stringResource(R.string.greeting_morning)
+        hour < 18 -> stringResource(R.string.greeting_afternoon)
+        else -> stringResource(R.string.greeting_evening)
     }
 }
-
-@Composable
-private fun HomeSectionContent(
-    section: HomeSection,
-    allCourses: List<Course>,
-    upcomingAssignments: List<Assignment>,
-    assignmentFilter: AssignmentFilter,
-    showIgnoredTab: Boolean,
-    ignoredAssignmentIds: Set<String>,
-    markedCompletedIds: Set<String>,
-    isLoggedIn: Boolean,
-    isLoading: Boolean,
-    initialLoadComplete: Boolean,
-    hasUnfinishedAssignment: (String) -> Boolean,
-    showAbsoluteTime: Boolean,
-    invertDirection: Boolean,
-    onCourseClick: (Course, String) -> Unit,
-    onAssignmentClick: (Assignment) -> Unit,
-    onToggleIgnore: (Assignment) -> Unit,
-    onMarkCompleted: (Assignment) -> Unit,
-    onSelectFilter: (AssignmentFilter) -> Unit,
-    onWidgetClick: () -> Unit,
-    onOpenSignInSettings: () -> Unit,
-) {
-    Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-        when (section.type) {
-            HomeSection.HomeSectionType.TODAY_COURSES -> {
-                TimeSliderSection(
-                    courses = allCourses,
-                    invertDirection = invertDirection,
-                    isLoggedIn = isLoggedIn,
-                    initialLoadComplete = initialLoadComplete,
-                    onSelectCourse = onCourseClick
-                )
-            }
-
-            HomeSection.HomeSectionType.UPCOMING_ASSIGNMENTS -> {
-                SectionHeader(
-                    title = if (section.type == HomeSection.HomeSectionType.CUSTOM) section.title else stringResource(
-                        section.type.defaultTitleRes
-                    )
-                )
-                if (isLoggedIn) {
-                    AssignmentFilterTabs(
-                        selected = assignmentFilter,
-                        enabled = true,
-                        showIgnoredTab = showIgnoredTab,
-                        onSelect = onSelectFilter,
-                    )
-                }
-                if (upcomingAssignments.isEmpty()) {
-                    AssignmentsEmptyState(
-                        isLoggedIn = isLoggedIn,
-                        isLoading = isLoading || (!initialLoadComplete && isLoggedIn),
-                        filter = assignmentFilter,
-                        onOpenSignInSettings = onOpenSignInSettings,
-                    )
-                } else {
-                    // Resolve the canonical Course for each row's courseNo so
-                    // the "name • courseNo" label reflects user renames and
-                    // the real NTUST code (iOS parity). Memoized to avoid
-                    // rebuilding the map on unrelated state changes.
-                    val courseByNo = remember(allCourses) {
-                        allCourses.associateBy { it.courseNo }
-                    }
-                    Card(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(horizontal = 16.dp),
-                        colors = CardDefaults.cardColors(
-                            containerColor = MaterialTheme.colorScheme.surface
-                        )
-                    ) {
-                        upcomingAssignments.forEachIndexed { index, assignment ->
-                            // SwipeableAssignmentRow keeps an Animatable keyed
-                            // by assignmentId; without `key()` here the parent
-                            // forEach has no item identity, so removing or
-                            // ignoring an assignment would rebind the next
-                            // row to the prior Animatable and leak swipe state.
-                            key(assignment.assignmentId) {
-                                val isMarkedCompleted =
-                                    assignment.assignmentId in markedCompletedIds
-                                SwipeableAssignmentRow(
-                                    assignment = assignment,
-                                    course = courseByNo[assignment.courseNo],
-                                    isIgnored = assignment.assignmentId in ignoredAssignmentIds,
-                                    isMarkedCompleted = isMarkedCompleted,
-                                    showAbsoluteTime = showAbsoluteTime,
-                                    onClick = { onAssignmentClick(assignment) },
-                                    onToggleIgnore = { onToggleIgnore(assignment) },
-                                    onMarkCompleted = { onMarkCompleted(assignment) },
-                                )
-                                if (index < upcomingAssignments.lastIndex) {
-                                    HorizontalDivider(modifier = Modifier.padding(horizontal = 16.dp))
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-
-            @Suppress("DEPRECATION")
-            HomeSection.HomeSectionType.QUICK_WIDGETS,
-            HomeSection.HomeSectionType.CUSTOM -> {
-                SectionHeader(
-                    title = if (section.type == HomeSection.HomeSectionType.CUSTOM) section.title else stringResource(
-                        section.type.defaultTitleRes
-                    )
-                )
-                if (section.widgets.isNotEmpty()) {
-                    LazyRow(
-                        contentPadding = PaddingValues(horizontal = 16.dp),
-                        horizontalArrangement = Arrangement.spacedBy(12.dp)
-                    ) {
-                        items(section.widgets, key = { it.id }) { widget ->
-                            Card(
-                                modifier = Modifier
-                                    .size(80.dp)
-                                    .clickable { onWidgetClick() },
-                                colors = CardDefaults.cardColors(
-                                    containerColor = MaterialTheme.colorScheme.surface
-                                )
-                            ) {
-                                Column(
-                                    modifier = Modifier.fillMaxSize(),
-                                    horizontalAlignment = Alignment.CenterHorizontally,
-                                    verticalArrangement = Arrangement.Center
-                                ) {
-                                    Icon(
-                                        imageVector = widget.feature.icon,
-                                        contentDescription = null,
-                                        tint = MaterialTheme.colorScheme.primary,
-                                        modifier = Modifier.size(28.dp)
-                                    )
-                                    Spacer(Modifier.height(4.dp))
-                                    Text(
-                                        text = stringResource(widget.feature.displayNameRes),
-                                        style = MaterialTheme.typography.labelSmall,
-                                        maxLines = 1
-                                    )
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
 @Composable
 private fun CourseDetailDialog(
     course: Course,
@@ -704,280 +432,6 @@ private fun CourseDetailDialog(
         },
     )
 }
-
-@Composable
-private fun greetingText(): String {
-    val hour = AppClock.calendar().get(Calendar.HOUR_OF_DAY)
-    return when {
-        hour < 6 -> stringResource(R.string.greeting_very_early)
-        hour < 12 -> stringResource(R.string.greeting_morning)
-        hour < 18 -> stringResource(R.string.greeting_afternoon)
-        else -> stringResource(R.string.greeting_evening)
-    }
-}
-
-private fun openAssignmentInMoodle(context: Context, assignment: Assignment) {
-    val targets = listOfNotNull(assignment.moodleDeepLink, assignment.moodleUrl)
-    for (target in targets) {
-        val intent = Intent(Intent.ACTION_VIEW, target.toUri()).apply {
-            addFlags(Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_NEW_DOCUMENT)
-        }
-        val opened = runCatching { context.startActivity(intent) }.isSuccess
-        if (opened) return
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun AssignmentFilterTabs(
-    selected: AssignmentFilter,
-    enabled: Boolean,
-    showIgnoredTab: Boolean,
-    onSelect: (AssignmentFilter) -> Unit,
-) {
-    // Hide 已忽略 when the user has nothing ignored — unless they're already
-    // on that tab, so clearing the last item doesn't yank the section out
-    // from under them mid-interaction.
-    val options = if (showIgnoredTab) AssignmentFilter.entries
-    else AssignmentFilter.entries.filter { it != AssignmentFilter.IGNORED }
-    val segmentColors = tigerDuckSegmentedButtonColors()
-    SingleChoiceSegmentedButtonRow(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 16.dp),
-    ) {
-        options.forEachIndexed { index, option ->
-            SegmentedButton(
-                selected = enabled && option == selected,
-                onClick = { onSelect(option) },
-                enabled = enabled,
-                shape = SegmentedButtonDefaults.itemShape(index = index, count = options.size),
-                colors = segmentColors,
-            ) {
-                Text(
-                    stringResource(option.displayNameRes),
-                    style = MaterialTheme.typography.labelMedium
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun tigerDuckSegmentedButtonColors(): SegmentedButtonColors {
-    val primary = MaterialTheme.colorScheme.primary
-    // Blend toward white in dark mode so the tint reads as a *lighter* pastel
-    // accent. A raw primary tint composites darker than the dark card surface.
-    val activeContainer = if (TigerDuckTheme.isDarkMode) {
-        lerp(primary, Color.White, 0.7f).copy(alpha = 0.22f)
-    } else {
-        primary.copy(alpha = 0.08f)
-    }
-    return SegmentedButtonDefaults.colors(
-        activeContainerColor = activeContainer,
-        activeContentColor = primary,
-        activeBorderColor = primary,
-    )
-}
-
-@Composable
-private fun AssignmentsEmptyState(
-    isLoggedIn: Boolean,
-    isLoading: Boolean,
-    filter: AssignmentFilter,
-    onOpenSignInSettings: () -> Unit = {},
-) {
-    if (!isLoggedIn) {
-        EmptyStateView(
-            icon = Icons.Filled.Lock,
-            title = stringResource(R.string.common_not_signed_in),
-            message = stringResource(R.string.common_sign_in_required_feature),
-            onIconClick = onOpenSignInSettings,
-        )
-        return
-    }
-    if (isLoading) {
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(32.dp),
-            contentAlignment = Alignment.Center,
-        ) {
-            CircularProgressIndicator(
-                modifier = Modifier.size(24.dp),
-                strokeWidth = 2.dp,
-            )
-        }
-        return
-    }
-    when (filter) {
-        AssignmentFilter.INCOMPLETE -> EmptyStateView(
-            icon = Icons.Filled.CheckCircle,
-            title = stringResource(R.string.home_assignments_all_good),
-            message = stringResource(R.string.home_assignments_none),
-        )
-
-        AssignmentFilter.ALL -> EmptyStateView(
-            icon = Icons.Filled.Inbox,
-            title = stringResource(R.string.home_assignments_none_now),
-            message = "",
-        )
-
-        AssignmentFilter.IGNORED -> EmptyStateView(
-            icon = Icons.Filled.VisibilityOff,
-            title = stringResource(R.string.home_assignments_no_ignored),
-            message = stringResource(R.string.home_assignments_ignore_hint),
-        )
-    }
-}
-
-/**
- * Assignment row with two swipe gestures, both following the SlotCard
- * pattern (0.6× drag damping, 100dp threshold, tween-out + snap-reset on
- * commit, spring-back otherwise):
- *  - Left swipe: toggle ignore (orange icon on the right edge — eye-off
- *    when not yet ignored, undo arrow when already ignored).
- *  - Right swipe: toggle mark-as-complete (green icon on the left edge —
- *    tick when not yet marked, undo arrow when already marked). The undo
- *    affordance is what makes 標示為完成 reversible from the 全部 tab.
- */
-@Composable
-private fun SwipeableAssignmentRow(
-    assignment: Assignment,
-    course: Course?,
-    isIgnored: Boolean,
-    isMarkedCompleted: Boolean,
-    showAbsoluteTime: Boolean,
-    onClick: () -> Unit,
-    onToggleIgnore: () -> Unit,
-    onMarkCompleted: () -> Unit,
-) {
-    val latestOnToggleIgnore by rememberUpdatedState(onToggleIgnore)
-    val latestOnMarkCompleted by rememberUpdatedState(onMarkCompleted)
-    val density = LocalDensity.current
-    val thresholdPx = with(density) { 100.dp.toPx() }
-    val swipeOffset = remember(assignment.assignmentId) { Animatable(0f) }
-    val coroutineScope = rememberCoroutineScope()
-    // Modifier.offset { ... } and Alignment.CenterStart/End are rtl-aware,
-    // but pointer deltas are raw screen-space. Negate in RTL so swipeOffset
-    // stays "logical" (positive = start) and the visual row tracks the finger.
-    val isRtl = LocalLayoutDirection.current == LayoutDirection.Rtl
-    val ignoreColor = Color(0xFF8E8E93)
-    val completeColor = Color(0xFF34C759)
-
-    Box(modifier = Modifier.fillMaxWidth()) {
-        val progress = (abs(swipeOffset.value) / thresholdPx).coerceIn(0f, 1f)
-
-        // Left-edge mark/unmark icon: revealed when the row is dragged
-        // right (positive offset). Tick when the row isn't marked yet,
-        // revert-arrow when it is already marked (so 標示為完成 is
-        // reversible without leaving the 全部 tab).
-        if (swipeOffset.value > 0f) {
-            Box(
-                modifier = Modifier
-                    .matchParentSize()
-                    .padding(start = 20.dp),
-                contentAlignment = Alignment.CenterStart,
-            ) {
-                Icon(
-                    imageVector = if (isMarkedCompleted) Icons.AutoMirrored.Filled.Undo
-                    else Icons.Filled.Check,
-                    contentDescription = if (isMarkedCompleted) stringResource(R.string.assignment_mark_complete_undo) else stringResource(
-                        R.string.assignment_mark_complete
-                    ),
-                    tint = completeColor,
-                    modifier = Modifier
-                        .size(26.dp)
-                        .alpha(progress)
-                        .scale(0.5f + 0.5f * progress),
-                )
-            }
-        }
-
-        // Right-edge hide/undo icon: revealed when the row is dragged left
-        // (negative offset).
-        if (swipeOffset.value < 0f) {
-            Box(
-                modifier = Modifier
-                    .matchParentSize()
-                    .padding(end = 20.dp),
-                contentAlignment = Alignment.CenterEnd,
-            ) {
-                Icon(
-                    imageVector = if (isIgnored) Icons.AutoMirrored.Filled.Undo
-                    else Icons.Filled.VisibilityOff,
-                    contentDescription = if (isIgnored) stringResource(R.string.assignment_ignore_undo) else stringResource(
-                        R.string.assignment_ignore
-                    ),
-                    tint = ignoreColor,
-                    modifier = Modifier
-                        .size(26.dp)
-                        .alpha(progress)
-                        .scale(0.5f + 0.5f * progress),
-                )
-            }
-        }
-
-        // Opaque row that slides over the indicators. Surface color matches
-        // the parent Card so no seam is visible at rest.
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .offset { IntOffset(swipeOffset.value.roundToInt(), 0) }
-                .background(MaterialTheme.colorScheme.surface)
-                .pointerInput(assignment.assignmentId) {
-                    detectHorizontalDragGestures(
-                        onDragEnd = {
-                            coroutineScope.launch {
-                                when {
-                                    swipeOffset.value <= -thresholdPx -> {
-                                        swipeOffset.animateTo(
-                                            -2000f,
-                                            animationSpec = tween(durationMillis = 200),
-                                        )
-                                        if (isRtl) latestOnMarkCompleted() else latestOnToggleIgnore()
-                                        swipeOffset.snapTo(0f)
-                                    }
-
-                                    swipeOffset.value >= thresholdPx -> {
-                                        swipeOffset.animateTo(
-                                            2000f,
-                                            animationSpec = tween(durationMillis = 200),
-                                        )
-                                        if (isRtl) latestOnToggleIgnore() else latestOnMarkCompleted()
-                                        swipeOffset.snapTo(0f)
-                                    }
-
-                                    else -> swipeOffset.animateTo(0f, animationSpec = spring())
-                                }
-                            }
-                        },
-                        onDragCancel = {
-                            coroutineScope.launch {
-                                swipeOffset.animateTo(0f, animationSpec = spring())
-                            }
-                        },
-                        onHorizontalDrag = { _, delta ->
-                            val signedDelta = if (isRtl) -delta else delta
-                            coroutineScope.launch {
-                                swipeOffset.snapTo(swipeOffset.value + signedDelta * 0.6f)
-                            }
-                        },
-                    )
-                },
-        ) {
-            AssignmentItem(
-                assignment = assignment,
-                course = course,
-                showAbsoluteTime = showAbsoluteTime,
-                markedCompleted = isMarkedCompleted,
-                isIgnored = isIgnored,
-                onClick = onClick,
-            )
-        }
-    }
-}
-
 @Composable
 private fun SyncConflictDialog(
     conflicts: List<HomeViewModel.SyncConflict>,
